青鑫：开始做时间流逝的逻辑

+++ Code/Prometheus/Assets/Scripts/Logical/Brick/BrickCore.cs
@@ -139,7 +139,7 @@ public class BrickCore : SingleObject<BrickCore> , IGetNode {
 
                         if (prob > Random.Range(0f, 1f))
                         {
-                            _brick = StageView.Instance.AddTablet(int.Parse(infos[1]));
+                            _brick = StageView.Instance.AddTablet(ulong.Parse(infos[1]));
                         }
                         else
                         {
